+++ src/php/Compiler/Domain/Parser/ParserNode/NumberNode.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Phel\Compiler\Domain\Parser\ParserNode;
 
 use Phel\Lang\BigDecimal;
-use Phel\Lang\BigInteger;
+use Phel\Lang\BigInt;
 use Phel\Lang\Rational;
 
 /**
- * @extends AbstractAtomNode<BigDecimal|BigInteger|float|int|Rational>
+ * @extends AbstractAtomNode<BigDecimal|BigInt|float|int|Rational>
  */
 final class NumberNode extends AbstractAtomNode {}
